Add workaround for highlighting matching brackets with matchparen (#85)
* Rename `tsxEscapeJs` syntax type to `tsxEscJs` as a work around to
mismatching bracket highlighting.
* Alias `tsxEscJs` to `tsxEscapeJs` to keep backward compatability for
styling.

The built in plugin matchparen highlights the matching paren when the
cursor is over a bracket. The matching relies on syntax types (when enabled)
and is configured to exclude any syntax types containing the words
"string\\|character\\|singlequote\\|escape\\|comment" from the matching
process.

This would cause a `tsxEscapeJs` region to mismatch highlighting for
that region and the overlapping regions bounded by braces.