Fix parenthesized type not being highlight inside arguments
Because the start pattern of typescriptFuncType group clashes with the
start pattern of typescriptParenthesizedType, the former would take
precedencec because it was declared last.

This tries to solve this issue by giving a more specific start pattern,
by assuming that it is only a typescriptFuncType if the parenthesis are
immediately followed by another parenthesis `)` or a set of keywords
followed by a colon `\k\+:`.

Hence the regex `\(\k\+:\|)\)\@=`.

Closes #233.